changed from float to double values

+++ src/vector/Vector2D.java
@@ -6,16 +6,16 @@ package vector;
 public class Vector2D {
 
     /**
-     * x and y coordinate stored as float
+     * x and y coordinate stored as double
      */
-    public float x, y;
+    public double x, y;
 
     /**
      * initialize 2D vector with a x and y coordinate
      * @param x coordinate
      * @param y coordinate
      */
-    public Vector2D(float x, float y) {
+    public Vector2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
